Module structures::list::linked_list

Structs

IntoIter
Iterator for lists by value.

Iter
Iterator for lists by reference.

Enums

LinkedList
A simple recursively defined linked list enumeration type.

Type Definitions

MoveResult
A convenience type for results which are used to move ownership back to the caller in the case of an error.