Enum structures::list::linked_list::LinkedList

pub enum LinkedList<T> {
    Cons(T, Box<LinkedList<T>>),
    Nil,
}

A simple recursively defined linked list enumeration type.

This implementation of a linked list is purely for the purpose of learning, and as reference of the details of a recursive type. You probably shouldn't use this data structure in real world applications.

Variants

Cons

Nil

Methods

impl<'a, T> LinkedList<T>

fn new() -> LinkedList<T>

Return a new empty linked list. This is semantically equivlent to List::Nil.

Examples

use structures::list::LinkedList;

assert_eq!(LinkedList::new(), LinkedList::Nil::<u32>);

fn is_empty(&self) -> bool

Determine if a linked list is empty.

Examples

use structures::list::LinkedList;

let list = LinkedList::<u32>::new();
assert!(list.is_empty());

fn len(&self) -> usize

Return the length of a list. This function relies on the implementation of into_iter() for creating a iterator over references to the elements.

Examples

use structures::list::LinkedList;

let list = LinkedList::new().push(1);
assert_eq!(list.len(), 1);

fn push(self, item: T) -> LinkedList<T>

Add an element to the front of the list. Insert is constant time because we return the new list containing the item, and the old now moved list.

Examples

use structures::list::LinkedList;

let list_a = LinkedList::new().push(2).push(1);
let list_b = list_a.push(0);

assert_eq!(list_b.len(), 3);

// Compiler error: use of moved value: `list_a`
// assert_eq!(list_a.len(), 3);

fn pop(self) -> Option<(T, LinkedList<T>)>

Return both the first and rest of a list, or None if called on Nil. Here Nil represents the end of the list, and contains no item. The returned values are moved, so it's important store them if they are needed later.

Examples

use structures::list::LinkedList;

let list = LinkedList::new().push("bye").push("hi");
let (first, rest) = list.pop().unwrap();

assert_eq!(first, "hi");
assert_eq!(rest, LinkedList::new().push("bye"));

// Compiler error: use of moved value: `list`
// assert_eq!(list.len(), 2);

fn insert(self, index: usize, item: T) -> MoveResult<LinkedList<T>, T>

Returns the list with an item inserted at the given index. Indexing starts at zero. For example (Cons 7 (Cons 9 (Cons 3))) has 7 at index 0, 9 at 1, and 3 at 2. Elements in the list after the inserted item will all have an index one larger than before, but the order of these elements is maintained.

This function returns a Result to handle the case when index is out of bounds. In this case, the original list is returned in the Err to allow for it's continued use.

Examples

use structures::list::LinkedList;

let list_a = LinkedList::new().push(0)
                              .push(1)
                              .push(2)
                              .insert(1, 42)
                              .unwrap();
let list_b = LinkedList::new().push(0)
                              .push(1)
                              .push(42)
                              .push(2);

assert_eq!(list_a, list_b);

fn remove(self, index: usize) -> MoveResult<(T, LinkedList<T>), T>

Removes an item by index in the list. Like pop, both the item and the rest of the list are moved and returned. Indexing starts at 0, see insert for an example of the indexing. Like insert the order of persisted elements is unchanged.

This function returns a Result to handle the case when index is out of bounds. In this case, the original list is returned in the Err to allow for it's continued use.

Examples

use structures::list::LinkedList;

let (item, list) = LinkedList::new().push(1)
                                    .push(2)
                                    .push(3)
                                    .remove(2)
                                    .unwrap();

assert_eq!(item, 1);
assert_eq!(list, LinkedList::new().push(2).push(3))