structures::list::linked_list::MoveResult

type MoveResult<T, U> = Result<T, LinkedList<U>>;

A convenience type for results which are used to move ownership back to the caller in the case of an error.

A simple example of this is the remove function which when called with an index that is out of bounds returns Err(original_list) thus transferring ownership. This allows for chaining functions, while still allowing to recover from errors easily.